Move CodeHighlightNode to core, move TableActionsPlugin back to playground
